requirements.txt
python-dotenv==1.2.1
browser-use==0.11.3
openai==2.15.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
